Merged PR 7658: Prograde: Upgrade Brine Properties
Brine description, UserDefined flag are updated. Resetting the FluidTypeIoTbl as per the referred brines in StratIoTbl.  HeatCaptype and ThermCondtype fields of FluidTypeIoTbl is updated as per the brine. Also FltThCondIoTbl and FltHeatCapIoTbl are updated to include the related values wrt all the referred ThermCondtype /HeatCaptype of  FluidTypeIoTbl.

Related work items: #314739, #323773