Merged PR 7942: Revert commit ded55286
Revert commit ded55286: This is the check regarding SIT environment issue. We don't have any other environment to check. We can't test this fix in the back-end side, because these scripts are not used if we run RT suit in the back-end.
This fix causes RT failures in Jenkins. So these changes are reverted back.