Merged PR 9388: Porosity models are updated from SoilMechanics to Exponential for all legacy UserDefined Lithologies
User defined lithologies with SM model is updated to Exponential model. The corresponding exponential model parameters are upadted based on the parent lithology name which is derived from the **DefinedBy** filed. If the parent lithology is not available there then it is extracted from the **Description** field. This is done by following the mapping sheet available for some selected lithologies available in PROD environment only.  Please refer to the PBI #351336 336 for more details

Related work items: #351336