Merged PR 9146: To check lithotypIoTbl field values are in BPA2 new value range
Io Table to work with: LithotypeIO
For User defined lithologies:

In LithotypeIO table under "DefinedBy" column:  Update this field with the name of the new Parent lithology determined by "BPA REF INFOI<This Name>|<Not this Name>".

Check whether the  property values for all the user defined lithologies of BPA1 are within the permissible limit of BPA2. Permissible limits of BPA2 lithological properties can be found in "Allowed Range for Lithotype Properties.xlsx" tab of the attached mapping sheet with the PBI. If any value is found to be outside the permissible limits of BPA2 then clip the values to the nearest limiting value (Need confirmations on the final ranges of BPA2 lithotype properties).

Related work items: #348585