Merged PR 9657: Bug fixed in lithology converter unit test case
After the review I changed the lithology test converter code (EXPECT_NEAR to EXPECT_EQ) and missed to remove the error value. This error was not captured while building in windows, but captured in linux build.
I modified the code and it is building fine in linux.

This pull request is showing the complete list of files. But the required change is only done in "LithologyConverterTest.cpp" file and "LithologyUpgradeManager.cpp" (changed std::double_t to double)

Related work items: #348585