Merged PR 10371: Updating the simulation window of legacy scenario if the legacy inputs are not meeting the BPA2 validation
A 3d scenario can be successfully imported to BPA2 if and only if, the simulation window defined has a minimum number of 3 nodes in both x-y direction. This also holds true for a scenario with sub-sampling size defined (the low-resolution grid). However, this validation of having 3 nodes in the simulation window minimum is not enforced in BPA-legacy and hence such scenarios will not be imported into BPA2 while bulk migration is performed. This implementation is to reset/increase the simulation window to have atleast 3 nodes in each direction. The strategy adopted here is:
1. First increase the WindowMax by keeping the windowMin and sub-sampling fixed. If the AOI is sufficiently large wrt to the new windowMax then reset the windowMax value.
2. If the newly calculated windowMax is going out of the defined AOI (specified by nodeX or nodeY), then try to reset the windowMin by keeping WindowMax and sub-sampling fixed.
3. if both steps 1 and 2 are goind out of the AOI, then reduce the sub-sampling size (if possible, it has to be greater than 1) and repeat the same step 1 and 2 until the new window comes within the AOI.

Related work items: #383678